Isolate all the casing functions to properly address the acronym casing.
